Fix breadcrumb link logic when moving past the last breadcrumb
Decidim's breadcrumbs stop at the component level, so if you reach a single proposal (show -view) it doesn't add the proposal anymore, the last breadcrumb will be the component. For this reason the old logic didn't apply a link in the component -breadcrumb because it 'read' it as the current page. A check has been added to not only check that the last breadcrumb is the last one, but also if it's the current page or not. If it's not the current page but last breadcrumb, it means we can enable the link for that breadcrumb since we are 1 step further from that page.